clear와 fastClear 통합 이미 점유하는 포트를 피해서 server를 실행하도록 수정

+++ index.js
@@ -13,13 +13,15 @@ class FakeIoRedis extends IoRedis {
     }
 
     static clear() {
-        server.kill();
-        server = new RedisServer();
-    }
-
-    static fastClear() {
         const cli = new FakeIoRedis(server.port);
-        return cli.flushall();
+        return cli.pubsub('channels', '*').then(r => {
+            if (r.length === 0) {
+                return cli.flushall();
+            }
+
+            server.kill();
+            server = new RedisServer();
+        });
     }
 }
